+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+- Add fallback functionality for Edge driver that in case of failure of downloading
+  auto/latest driver tries to download driver of the previous version.
+
+### Changed
+
+- Improve error handling of driver downloading.
 
 ## [2.11.0] - 2024-07-18