+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.12.0] - 2024-07-30
+
 ### Added
 
-- Add fallback functionality for Edge driver that in case of failure of downloading
-  auto/latest driver tries to download driver of the previous version.
+- Add fallback functionality for Edge driver that in case of failure of downloading auto/latest driver tries to download driver of the previous version.
+  Fixes issue [#16 Edge web driver creation fails with "End of Central Directory record could not be found."](https://github.com/atata-framework/atata-webdriversetup/issues/16).
 
 ### Changed
 
@@ -145,3 +147,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2021-02-16
 
 Initial version release.
+
+[Unreleased]: https://github.com/atata-framework/atata-templates/compare/v2.12.0...HEAD
+[2.12.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.11.0...v2.12.0
+[2.11.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.10.0...v2.11.0
+[2.10.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.9.0...v2.10.0
+[2.9.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.8.0...v2.9.0
+[2.8.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.7.0...v2.8.0
+[2.7.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.6.0...v2.7.0
+[2.6.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.5.0...v2.6.0
+[2.5.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.4.0...v2.5.0
+[2.4.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.3.0...v2.4.0
+[2.3.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.2.0...v2.3.0
+[2.2.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.1.0...v2.2.0
+[2.1.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v2.0.0...v2.1.0
+[2.0.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v1.2.1...v2.0.0
+[1.2.1]: https://github.com/atata-framework/atata-webdriversetup/compare/v1.2.0...v1.2.1
+[1.2.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v1.1.0...v1.2.0
+[1.1.0]: https://github.com/atata-framework/atata-webdriversetup/compare/v1.0.0...v1.1.0
+[1.0.0]: https://github.com/atata-framework/atata-webdriversetup/releases/tag/v1.0.0